Rename updatedColumns to manualAnnotations
Make this a dictionary with a field for the columns. It was weird to
have the array of columns wrapped in a dict, but giving this a more
general name allows us to put other information in there (such as
pairing for coverage) later.

Also don't store it as JSON string but as an object.