Limit the number of Lazo candidates sent to ES
Elasticsearch has a static limit (from server configuration) for the
number of clauses in a boolean query. If we have more results from Lazo,
we can't send them all or we'll get a 400 from Elasticsearch.

This means not all Lazo results are considered, but it's better than
having the search return an error.