Prevent asyncio tasks from being destroyed
On Python 3.8, I am seeing "Task was destroyed but it is pending!". This
is probably caused by the log_future() calls, which are leaking the Task
on purpose. This commit keeps a strong reference to the Task's future
until it resolves, which should prevent the warning.

See also https://bugs.python.org/issue21163